feat : query 디렉토리로 이동 및 평탄화
기존 방식으로 Query 결과를 담으려면 Json 처리가 필수적, 이에 따라 평탄화 이후 원래 DTO로 변환하는 과정 추가
